Create vehicles model
